fix: change received field to integer (#72)
In sample format v1 and legacy android profiles, `received` is either
expected to be either a Time "parsable" (rfc 3339 formatted string) or
an int.

Therefore we're changing the type to int in V1 implementation to make it
compatible with what vroom expect.

+++ src/android/profile.rs
@@ -68,7 +68,7 @@ pub struct AndroidProfile {
 
     project_id: u64,
 
-    received: f64,
+    received: i64,
 
     release: Option<String>,
 
@@ -134,7 +134,7 @@ impl ProfileInterface for AndroidProfile {
         self.project_id
     }
 
-    fn get_received(&self) -> f64 {
+    fn get_received(&self) -> i64 {
         self.received
     }
